Fix: Clerk Webhook API for user creation and user deletion

+++ app/server/user.action.ts
@@ -16,13 +16,24 @@ export const createUser = async (user: UserType) => {
 
 export const deleteUser = async (clerkUserID: string) => {
     try {
+        const userId = await prisma.user.findFirst({ where: { clerkUserID: clerkUserID } }).then(user => user?.id);
+
+        if (!userId)
+            throw new Error("User not found");
+
+        await prisma.foodItemList.deleteMany({
+            where: {
+                user_id: userId,
+            }
+        });
+
         await prisma.user.delete({
             where: {
-                clerkUserID: clerkUserID,
+                id: userId,
             }
         });
 
     } catch (error) {
-        console.error(error);
+        console.error("Error deleting user: " + ": " + error);
     }
 }
